Move TestMemCumeDist Func to func_cume_dist_test.go
